Extract fetch implementation to its own package
