+++ packages/kubernetes-client-angular/package.json
@@ -29,7 +29,6 @@
     "@angular/platform-browser": "^15.2.0",
     "@angular/platform-browser-dynamic": "^15.2.0",
     "@angular/router": "^15.2.0",
-    "@ccremer/kubernetes-client": "^0.5.0",
     "@ngrx/data": "^15.4.0",
     "@ngrx/effects": "^15.4.0",
     "@ngrx/store": "^15.4.0",
@@ -38,6 +37,9 @@
     "rxjs": "~7.8.0",
     "tslib": "2.5.0",
     "zone.js": "~0.13.0"
+  },
+  "peerDependencies": {
+    "@ccremer/kubernetes-client": "*"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^15.2.6",
