+++ packages/kubernetes-client-example-fetch/src/js/main.ts
@@ -1,8 +1,8 @@
 import '../styles.scss'
-import { Client, KubeClientBuilder } from '@ccremer/kubernetes-client/fetch'
+import { Client, KubeClientBuilder } from '../../../kubernetes-client-fetch/src'
 import { newSelfSubjectRulesReview } from './types'
 import { createAlert } from './alerts'
-import { WatchEvent } from '@ccremer/kubernetes-client/api'
+import { WatchEvent } from '@ccremer/kubernetes-client-fetch'
 import { KubeObject } from '@ccremer/kubernetes-client/types/core'
 
 console.debug('Starting up...')
