+++ packages/kubernetes-client-fetch/.eslintrc.json
@@ -1,0 +1,4 @@
+{
+  "root": true,
+  "extends": ["custom"]
+}
