+++ packages/kubernetes-client-fetch/.prettierignore
@@ -1,0 +1,3 @@
+# Ignore artifacts
+/dist/
+package-lock.json
