+++ packages/kubernetes-client-fetch/examples/custom-type.ts
