+++ packages/kubernetes-client-fetch/examples/vite.config.ts
