+++ packages/kubernetes-client-fetch/package.json
@@ -1,15 +1,14 @@
 {
-  "name": "@ccremer/kubernetes-client",
-  "version": "0.5.0",
+  "name": "@ccremer/kubernetes-client-fetch",
+  "version": "0.1.0",
   "description": "Generic Kubernetes client run in the Browser",
   "scripts": {
     "preformat": "node ../../README.cjs README.md.njk > README.md",
     "format": "prettier --write ./**/*.{js,ts,json}",
     "lint": "eslint './**/*.{js,ts}'",
     "prelint": "npm run format",
-    "test": "vitest run",
-    "clean": "rm -rf types fetch api index.*",
-    "build": "rm -rf dist && tsc",
+    "prebuild": "rm -rf dist",
+    "build": "tsc",
     "watch": "tsc --watch",
     "prepack": "cp package.json README.md dist/"
   },
@@ -33,13 +32,15 @@
   "types": "index.d.ts",
   "type": "module",
   "files": [
-    "api",
-    "types",
-    "index.*"
+    "*.d.ts",
+    "*.js"
   ],
   "devDependencies": {
     "eslint-config-custom": "*",
     "vitest": "0.30.1"
+  },
+  "peerDependencies": {
+    "@ccremer/kubernetes-client": "*"
   },
   "publishConfig": {
     "access": "public"
