+++ packages/kubernetes-client-fetch/src/index.ts
@@ -1,6 +1,6 @@
-export * from './client'
+export * from './authorizer'
 export * from './builder'
 export * from './client'
-export * from './authorizer'
+export * from './fetch-client'
 export * from './jsonlinestream'
 export * from './watcheventstream'
