+++ packages/kubernetes-client-fetch/src/jsonlinestream.ts
