+++ packages/kubernetes-client/src/api/index.ts
@@ -1,5 +1,4 @@
 export * from './options'
 export * from './config'
-export * from './client'
 export * from './pluralizer'
 export * from './urlgenerator'
